requirements.txt
aniso8601==8.0.0
asn1crypto==1.0.0
astroid==2.2.5
brindexer1==0.1
brindexer2==0.1
brindexer3==0.1
brindexer32==0.1
cds-brindexer==0.1
cds-flash-maintenance==0.1
certifi==2018.10.15
cffi==1.12.3
cleancat==0.7.8
Click==7.0
cryptography==2.7
dj-database-url==0.5.0
Django==2.1.2
django-tastypie==0.14.2
djangorestframework==3.11.0
Flask==1.1.1
Flask-API==1.1
Flask-Ext==0.1
flask-mongoengine==0.9.5
Flask-MongoRest==0.2.3
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
Flask-Views==0.2.1
Flask-WTF==0.14.2
gunicorn==20.0.4
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
kafka==1.3.5
kafka-python==1.4.6
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mimerender==0.6.0
mongoengine==0.18.2
mysql==0.0.2
mysqlclient==1.4.4
pipenv==2018.10.13
psycopg2==2.8.5
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pylint==2.3.1
pymongo==3.8.0
PyMySQL==0.9.3
pytest==3.0.7
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2019.2
six==1.12.0
test==0.1
test-brindexer==0.1
typed-ast==1.4.0
virtualenv==16.0.0
virtualenv-clone==0.4.0
Werkzeug==0.16.0
whitenoise==5.1.0
wrapt==1.11.1
WTForms==2.2.1
